requirements.txt
Flask==2.2.3
Flask-Session==0.4.0
pandas==1.3.5
openpyxl==3.1.2
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.76.0
werkzeug==2.2.3
numpy<2
numexpr==2.8.1
